+++ invisibleroads_scaffolds/__init__.py
@@ -1,0 +1,7 @@
+from pyramid.paster import PyramidTemplate
+
+
+class CoreTemplate(PyramidTemplate):
+
+    _template_dir = 'core'
+    summary = 'Pyramid SQLAlchemy project with user account management'