+++ invisibleroads_scaffolds/core/utilities/clear.sh
@@ -1,0 +1,1 @@
+rm -Rf data *.db *.db-journal logs/*.log