+++ invisibleroads_scaffolds/core/utilities/system/stop.sh
@@ -1,0 +1,3 @@
+for pid in `ps x | grep production.ini | grep -v grep | awk '{print $1}'`; do
+    kill $pid
+done